Move logging to its own directory
